requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.104.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
twilio==8.10.1
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
yarl==1.9.2
